pyproject.toml
[project]
name = "financial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mplfinance>=0.12.10b0",
    "pandas-ta>=0.4.71b0",
    "ta>=0.11.0",
    "yfinance>=0.2.66",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.13.3",
]
